refactor the home page to use tab enum; reuse translator for tab titles
