+++ lib/widgets/app.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:weforza/generated/l10n.dart';
-import 'package:weforza/widgets/pages/home_page.dart';
+import 'package:weforza/widgets/pages/home_page/home_page.dart';
 import 'package:weforza/widgets/platform/platform_aware_widget.dart';
 import 'package:weforza/widgets/theme.dart';
